refactor(simplygo): organise simplygo code into simplygo module
